creating test validation and json with user data
