+++ tests/cypress/support/commands.js
@@ -1,92 +1,3 @@
-// ***********************************************
-// This example commands.js shows you how to
-// create various custom commands and overwrite
-// existing commands.
-//
-// For more comprehensive examples of custom
-// commands please read more here:
-// https://on.cypress.io/custom-commands
-// ***********************************************
-//
-//
-// -- This is a parent command --
-// Cypress.Commands.add('login', (email, password) => { ... })
-//
-//
-// -- This is a child command --
-// Cypress.Commands.add('drag', { prevSubject: 'element'}, (subject, options) => { ... })
-//
-//
-// -- This is a dual command --
-// Cypress.Commands.add('dismiss', { prevSubject: 'optional'}, (subject, options) => { ... })
-//
-//
-// -- This will overwrite an existing command --
-// Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
 
-import loginPage from './pages/Login'
-import mapPage from './pages/Map'
-
-Cypress.Commands.add('apiLogin', (user) => {
-    const payload = {
-        instagram: user.instagram,
-        password: user.password
-    }
-
-    cy.request({
-        url: 'http://localhost:3333/sessions',
-        method: 'POST',
-        body: payload
-    }).then(response => {
-        expect(response.status).to.eql(200)
-        Cypress.env('token', response.body.token)
-    })
-})
-
-Cypress.Commands.add('apiResetUser', (instagram) => {
-    cy.request({
-        url: 'http://localhost:3333/helpers-reset',
-        method: 'DELETE',
-        qs: { instagram: instagram }
-    }).then(response => {
-        expect(response.status).to.eql(204)
-    })
-})
-
-Cypress.Commands.add('apiCreateUser', (payload) => {
-    cy.apiResetUser(payload.instagram)
-
-    cy.request({
-        url: 'http://localhost:3333/signup',
-        method: 'POST',
-        body: payload
-    }).then(response => {
-        expect(response.status).to.eql(201)
-    })
-})
-
-Cypress.Commands.add('apiCreateFoodTruck', (payload) => {
-    cy.request({
-        url: 'http://localhost:3333/foodtrucks',
-        method: 'POST',
-        headers: {
-            'Authorization': Cypress.env('token')
-        },
-        body: payload
-    }).then(response => {
-        expect(response.status).to.eql(201)
-    })
-})
-
-Cypress.Commands.add('uiLogin', (user) => {
-    loginPage.go('-23.632061492457712', '-46.71051979064942')
-    loginPage.form(user)
-    loginPage.submit()
-
-    mapPage.loggedUser(user.name)
-})
-
-Cypress.Commands.add('setGeolocation', (lat, long) => {
-    localStorage.setItem('qtruck:latitude', lat)
-    localStorage.setItem('qtruck:longitude', long)
-})
+import './commands/api'
+import './commands/ui'
